Add an alias to checkout the default branch
I use repos with a variety of default branches some of which have a main/master
branch which is not the default instead develop/dev is. This can  lead to
confusion when creating new branches to work on.

This alias checks out the default branch for a repo.

Inspired by https://stackoverflow.com/a/61357104